// types/vapiAi/api/calls/create.ts
export type CallStatus =
	| "queued"
	| "ringing"
	| "in-progress"
	| "forwarding"
	| "ended";